fix: type for rate limit in price calculator

+++ src/IPriceCalculator.sol
@@ -6,5 +6,5 @@ interface IPriceCalculator {
     /// @param _rateLimit the rate limit the user wants to acquire
     /// @return address of the erc20 token
     /// @return uint price to pay for acquiring the specified `_rateLimit`
-    function calculate(uint256 _rateLimit) external view returns (address, uint256);
+    function calculate(uint32 _rateLimit) external view returns (address, uint256);
 }
